requirements.txt
tqdm==4.65.0
numpy==1.25.1
transformers==4.26.1
torchtext==0.14.1  # this will automatically install torch==1.13.1 as its dependancy!
torchdata==0.5.1
mosestokenizer==1.2.1
wandb==0.15.7
st-annotated-text
flask
flask-cors
pynif